move keccak lib function-> lib folder
